functions/paginator/requirements.txt
appwrite
python-dotenv
mock
